requirements.txt
# Core dependencies
backboard-sdk>=1.0.0
cerebras-cloud-sdk>=1.0.0
httpx>=0.25.0
rich>=13.0.0
click>=8.0.0
python-dotenv>=1.0.0

# HTTP API (FastAPI)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# MCP Server
mcp>=1.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
mypy>=1.0.0
ruff>=0.1.0
